Cleanup to make gcc happy
std=gnu11 -Wpedantic -Werror

The more stringent Options did find a suspicious pointer from int where we
now up-cast prior function call. It's a convenience passing however to trick
pthread_create in accepting a pass-by-value for it's argument. It will work
as long as pointers are bigger or equal in size than int however.

To make sure, we've added an extra assert in main().