Legacy lib dependency removed (mqueue & log)
Builds and produces valid binaries for Android using NDK.

This project now has no source dependencies and can be built
without `git submodule sync`.

Historic recap:
==============

tcp-tap war originally written using mqueues for the switch-board
plumbing. This was later replaced by named pipes (fifos).

The implementation lib/mqueue is quite limited in queue-sizes as
it was primarily aiming for deeply embedded systems. This is now
vastly improved. mqueue is likely much faster but the fifos
do the job (functionally) just as well for now.

Minor traces of this remained which made it not possible to
build for Android without also pulling in submodule lib/mqueue,
which in turn has a dependency to submodule lib/log.

I.e. by eliminating dependency to mqueue, log can also be
removed.

**This also makes Android porting much easier** (even though
porting was already just completed).

Both modules are kept in source as gitmodules for reference.
Especially lib/log might come in handy as it has the same API as
Android LOG.