accept(): uninitialized addrlen (bug-fix)
From documentation on accept(2):

The addrlen argument is a value-result argument: it should
initially contain the size of the structure pointed to by addr;
on return it will contain the actual length (in bytes) of the
address returned. When addr is NULL nothing is filled in.

( This bug has been there for almost a decade :-) )