Use iatty() to track where TTY is lost
Loosing TTY can be due to either property of pipe when created, or
because neither stdio in process fork/execute tcp-tap was a connected to
a terminal in the first place. Not being connected to a terminal affects
wrapped programs like GDB in that their readln support is not
functionning.